+++ src/test/java/shopping/domain/cart/CartProductTest.java
@@ -1,10 +1,9 @@
-package shopping.domain;
+package shopping.domain.cart;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import shopping.domain.cart.CartProduct;
 
 @DisplayName("장바구니 상품 테스트")
 class CartProductTest {
