+++ src/test/java/shopping/domain/member/MemberPasswordTest.java
@@ -1,4 +1,4 @@
-package shopping.domain;
+package shopping.domain.member;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -7,7 +7,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-import shopping.domain.member.MemberPassword;
 import shopping.exception.ShoppingException;
 
 @DisplayName("회원 비밀번호 테스트")
